[Chore] #25-Extension 파일 수정(NSObject)
