--- HARA/HARA/Global/Extension/ClassIdentifier+.swift
@@ -1,27 +1,0 @@
-//
-//  ClassIdentifier+.swift
-//  HARA
-//
-//  Created by saint on 2023/01/10.
-//
-
-import Foundation
-import UIKit
-
-/**
-
-  - Description:
- 
-          각 VC,TVC,CVC의 className을 String으로 가져올 수 있도록 연산 프로퍼티를 설정합니다.
-          요 값들은 나중에 Identifier에 잘 써먹을 수 있습니다 ^__^
- */
-
-extension NSObject {
-  
-    static var classIdentifier: String {
-        NSStringFromClass(self.classForCoder()).components(separatedBy: ".").last!
-    }
-     var classIdentifier: String {
-        NSStringFromClass(self.classForCoder).components(separatedBy: ".").last!
-    }
-}
